refactor: move, rename, delete hooks & components
