+++ src/app/(app)/(map)/_components/List.tsx
@@ -3,10 +3,8 @@ import { T } from "@transifex/react";
 import { CircleAlert } from "lucide-react";
 import { useEffect, useMemo, useState } from "react";
 import styled from "styled-components";
+import { ListItem } from "~/app/(app)/(map)/_components/ListItem";
 import { SecondaryButton } from "~/components/button/SecondaryButton";
-import type { RenderedFeature } from "~/layouts/DefaultLayout";
-import { ListItem } from "~/modules/list/components/ListItem";
-import { useMap } from "~/modules/map/hooks/useMap";
 import { useRenderedFeatures } from "~/modules/map/hooks/useRenderedFeatures";
 import { getExternalSources } from "~/modules/map/utils/sources";
